Added a testing flag for the test suite, so we can savely do some unit testing against our code.
Added more logging options, so we can completely kill the logger if needed
or work with the emitted events of the logger to see if our code behaves as
expected.

Added an array of paths that we will scan for plugins, so it's easier later
on to fully implement a custom plugin interface for users..

Adjusted the stdout log silencer to use the Log levels from the Logger instance
to set the correct log level, as we don't want to output debug logs by default.

Started more work on our unit tests, as I don't want to introduce any regressions
after I have released the 1.0 (stable) release.